@@ -10,35 +10,36 @@
     "postinstall": "nuxt prepare"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.5.1",
-    "@fortawesome/vue-fontawesome": "^3.0.6",
-    "@imgly/background-removal": "^1.4.4",
-    "@nuxt/image": "^1.3.0",
-    "@pinia/nuxt": "^0.5.1",
+    "@fortawesome/fontawesome-svg-core": "^7.1.0",
+    "@fortawesome/vue-fontawesome": "^3.1.2",
+    "@imgly/background-removal": "^1.7.0",
+    "@nuxt/image": "^2.0.0",
+    "@pinia/nuxt": "^0.11.3",
     "@types/blueimp-load-image": "^5.16.6",
-    "@vueuse/nuxt": "^10.7.2",
+    "@vueuse/nuxt": "^14.0.0",
     "blueimp-load-image": "^5.16.0",
     "lodash-es": "^4.17.21",
-    "nuxt": "^3.10.2",
-    "vue": "^3.4.19",
+    "nuxt": "^4.2.1",
+    "pinia": "^3.0.4",
+    "vue": "^3.5.13",
     "vue-confetti-explosion": "^1.0.2",
-    "vue-router": "^4.2.5",
+    "vue-router": "^4.6.3",
     "vue3-roulette": "^0.3.3"
   },
   "devDependencies": {
-    "@element-plus/nuxt": "^1.0.7",
-    "@fortawesome/free-brands-svg-icons": "^6.5.1",
-    "@fortawesome/free-solid-svg-icons": "^6.5.1",
-    "@nuxt/test-utils": "^3.11.0",
-    "@nuxtjs/supabase": "^1.1.6",
-    "@nuxtjs/tailwindcss": "^6.11.4",
-    "@vue/test-utils": "^2.4.4",
-    "bun": "^1.0.28",
-    "element-plus": "^2.5.5",
-    "happy-dom": "^13.3.8",
-    "playwright-core": "^1.41.2",
-    "sass": "^1.70.0",
-    "sass-loader": "^14.1.0",
-    "vitest": "^1.3.1"
+    "@element-plus/nuxt": "^1.1.4",
+    "@fortawesome/free-brands-svg-icons": "^7.1.0",
+    "@fortawesome/free-solid-svg-icons": "^7.1.0",
+    "@nuxt/test-utils": "^3.20.1",
+    "@nuxtjs/supabase": "^2.0.1",
+    "@nuxtjs/tailwindcss": "^6.14.0",
+    "@vue/test-utils": "^2.4.6",
+    "bun": "^1.3.3",
+    "element-plus": "^2.11.8",
+    "happy-dom": "^20.0.10",
+    "playwright-core": "^1.57.0",
+    "sass": "^1.94.2",
+    "sass-loader": "^16.0.6",
+    "vitest": "^3.2.4"
   }
 }
