@@ -1,5 +1,6 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
+  compatibilityDate: '2025-11-25',
   devtools: { enabled: true },
   modules: [
     "@nuxt/test-utils/module",
@@ -54,7 +55,7 @@ export default defineNuxtConfig({
       login: "/login",
       callback: "/confirm",
       exclude: [],
-      cookieRedirect: true,
+      saveRedirectToCookie: true,
     },
   },
 });
